zenbeats-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
scikit-learn==1.5.0
diffusers==0.27.2
transformers==4.43.0
torch==2.2.2
accelerate==0.30.1
safetensors==0.4.2
Pillow==10.3.0